Pr fix

+++ backend/service/exam_srv.go
@@ -3,21 +3,24 @@ package service
 import (
 	"context"
 	"fmt"
+	"strings"
 
+	"latih.in-be/helper"
 	"latih.in-be/model"
 	"latih.in-be/repository"
 )
 
 type ExamService interface {
-	Create(ctx context.Context, data model.Exam) (*model.Exam, error)
+	Create(ctx context.Context, data model.Exam) error
 	GetById(ctx context.Context, id int) (*model.Exam, error)
-	Update(ctx context.Context, data model.Exam) (*model.Exam, error)
-	Delete(ctx context.Context, id int) error
+	Update(ctx context.Context, data model.Exam, id int) (*model.Exam, error)
+	Delete(ctx context.Context, id int, userId int) error
 	GetAll(ctx context.Context) ([]model.Exam, error)
 }
 
 type examService struct {
-	repo repository.ExamRepository
+	repo     repository.ExamRepository
+	userRepo repository.UserRepository
 }
 
 func NewExamService(repo repository.ExamRepository) ExamService {
@@ -26,21 +29,28 @@ func NewExamService(repo repository.ExamRepository) ExamService {
 	}
 }
 
-func (s *examService) Create(ctx context.Context, data model.Exam) (*model.Exam, error) {
+func (s *examService) Create(ctx context.Context, data model.Exam) error {
 	if data.CreatorId == 0 {
-		return nil, fmt.Errorf("creatorId is required")
+		return fmt.Errorf("creator_id is required")
 	}
 
-	if data.EndedAt.Before(data.StartedAt) {
-		return nil, fmt.Errorf("ended_at must be after started_at")
+	if data.StartedAt == nil {
+		return fmt.Errorf("started_at is required")
 	}
 
-	createdData, err := s.repo.Create(ctx, data)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create exam: %w", err)
+	if data.FinishedAt == nil {
+		return fmt.Errorf("finished_at is required")
 	}
 
-	return createdData, nil
+	if data.FinishedAt.Before(*data.StartedAt) {
+		return fmt.Errorf("finished_at must be after started_at")
+	}
+
+	if err := s.repo.Create(ctx, data); err != nil {
+		return fmt.Errorf("failed to create exam: %w", err)
+	}
+
+	return nil
 }
 
 func (s *examService) GetById(ctx context.Context, id int) (*model.Exam, error) {
@@ -51,11 +61,24 @@ func (s *examService) GetById(ctx context.Context, id int) (*model.Exam, error) 
 	return data, nil
 }
 
-func (s *examService) Update(ctx context.Context, data model.Exam) (*model.Exam, error) {
-	updatedData, err := s.repo.Update(ctx, data)
+func (s *examService) Update(ctx context.Context, data model.Exam, id int) (*model.Exam, error) {
+	updatedData, err := s.repo.Update(ctx, data, id)
 	if err != nil {
-		return nil, fmt.Errorf("update failed: %w", err)
+		if strings.Contains(err.Error(), "Unknown column") {
+			var fieldName string
+			parts := strings.Split(err.Error(), "'")
+			if len(parts) >= 2 {
+				fieldName = parts[1]
+			}
+
+			val := helper.GetFieldValue(data, fieldName)
+
+			return nil, fmt.Errorf("field '%s' with value '%v' is undefined", fieldName, val)
+		}
+
+		return nil, fmt.Errorf("update gagal: %v", err)
 	}
+
 	return updatedData, nil
 }
 
@@ -67,7 +90,21 @@ func (s *examService) GetAll(ctx context.Context) ([]model.Exam, error) {
 	return data, nil
 }
 
-func (s *examService) Delete(ctx context.Context, id int) error {
+func (s *examService) Delete(ctx context.Context, id int, userId int) error {
+	data, err := s.repo.GetById(ctx, id)
+	if err != nil {
+		return fmt.Errorf("data is unavaible %w", err)
+	}
+
+	user, err := s.userRepo.GetById(ctx, userId)
+	if err != nil {
+		return fmt.Errorf("user is unavaible %w", err)
+	}
+
+	if user.Id != data.CreatorId && user.Role != model.RoleAdmin {
+		return fmt.Errorf("you are not the creator or admin")
+	}
+
 	if err := s.repo.Delete(ctx, id); err != nil {
 		return fmt.Errorf("failed to delete data: %w", err)
 	}
